Refactor BinaryNoGoodPropagationEstimation
addressing https://github.com/alpha-asp/Alpha/pull/157#discussion_r270227739 by @AntoniusW
